// src/views/pages/authentication/types.ts
/**
 **********************************************************************
 * This source code is shared for the sole purpose of demo. Do not copy!
 **********************************************************************
 **/

export type OnboardingRegisterFormData = {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  aboutRemoved: string;
  industry: string;
};

export type PlanTypes = 'month' | 'year';